requirements.txt
# RAG Chatbot System Requirements

# Core dependencies for the RAG chatbot
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13

# Vector database and embeddings
faiss-cpu==1.7.4
numpy==1.24.3

# Azure OpenAI
openai==1.12.0

# Environment management
python-dotenv==1.0.0

# Data handling
pandas==2.0.3

# Web interface (optional)
streamlit==1.29.0
gradio==4.12.0

# Development and testing
pytest==7.4.0
jupyter==1.0.0

# Additional utilities
tqdm==4.66.1
rich==13.7.0
